backend/pyproject.toml
[project]
name = "hackathon-googlemaps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aci-sdk>=1.0.0b3",
    "black>=25.1.0",
    "elevenlabs>=2.7.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
